src/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
Pygments==2.14.0
pytz==2022.7.1
sqlparse==0.4.3
whitenoise==6.3.0
